Move CoinService to 'application' from 'services'
