+++ src/coin/application/coin_service.go
@@ -1,4 +1,4 @@
-package services
+package application
 
 import (
 	"coin/domain"
