+++ src/coin/infrastructure/integrations/bitfinex_client.go
@@ -1,7 +1,7 @@
 package integrations
 
 import (
-	"coin/application/services"
+	"coin/application"
 	"coin/domain"
 	"coin/infrastructure"
 	"encoding/json"
@@ -22,7 +22,7 @@ func NewBitfinexClient(client infrastructure.IHttpClient) *BitfinexClient {
 	return &BitfinexClient{client: client, baseURL: baseUrl}
 }
 
-func (b *BitfinexClient) GetRate(currency string, coin string) (*services.SpotPrice, error) {
+func (b *BitfinexClient) GetRate(currency string, coin string) (*application.SpotPrice, error) {
 	path := fmt.Sprintf("/pubticker/%s%s", coin, currency)
 	url := b.baseURL.String() + path
 
@@ -49,7 +49,7 @@ func (b *BitfinexClient) GetRate(currency string, coin string) (*services.SpotPr
 		return nil, err
 	}
 
-	return &services.SpotPrice{Amount: price, Timestamp: timestamp}, nil
+	return &application.SpotPrice{Amount: price, Timestamp: timestamp}, nil
 }
 
 type bitfinexResponse struct {
